+++ Toggly.FeatureManagement.Next/nextjs-toggly-server/src/index.ts
@@ -2,6 +2,7 @@
 export {
   initServerToggly,
   getServerToggly,
+  waitForServerToggly,
   useServerToggly,
   refreshServerToggly,
   isServerFeatureOn,